+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 4.23
 
+* Contribute Language Model Tools for Copilot Agent. Currently available tools are:
+  * Help setting up Connected Mode
+  * Exclude given file or folder from local analysis
+  * List Security Hotspots detected in the file
 * Update Go analyzer 1.22.1 -> [1.23](https://sonarsource.atlassian.net/issues/?jql=fixVersion%20=%2019232%20ORDER%20BY%20created%20ASC) -> [1.23.1](https://sonarsource.atlassian.net/issues/?jql=fixVersion%20%3D%2019661%20ORDER%20BY%20created%20ASC)
 * Update IaC analyzer 1.46 -> [1.47](https://sonarsource.atlassian.net/issues/?jql=fixVersion%20=%2019100%20ORDER%20BY%20created%20ASC)
 * Update PHP analyzer 3.45 -> [3.46](https://sonarsource.atlassian.net/issues/?jql=fixVersion%20%3D%2017781%20ORDER%20BY%20created%20ASC)